bp_ecg_etl/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
aioboto3==15.0.0
aiobotocore==2.23.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
attrs==25.3.0
aws-lambda-typing==2.20.0
boto3==1.38.27
botocore==1.38.27
frozenlist==1.7.0
idna==3.10
jmespath==1.0.1
multidict==6.6.3
pillow==11.3.0
polars==1.31.0
propcache==0.3.2
pymupdf==1.26.3
python-dateutil==2.9.0.post0
s3transfer==0.13.1
six==1.17.0
structlog==25.4.0
tenacity==9.1.2
typing-extensions==4.14.1 ; python_full_version < '3.13'
urllib3==2.5.0
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.1
